# [4.x] [PoC] Optimize vertex shader code register usage