Had some spare time today and decided to go over the vertex shader's ISA line by line while looking for optimization opportunities. I managed to reduce the needed registers from 71 to 53, which (in theory) should help us improve occupancy, which would in turn make it easier to cover up any waits in the vertex shader.

In my own testing I couldn't measure a noticeable difference, but that's mostly down to the fact that nvidia nsight is barely usable on linux. The performance numbers varied by up to 30% between different runs and sometimes were an order of magnitude off from the estimates I got with a simple in editor script. Part of this is probably caused by the fact that nsight really dislikes how we interleave our frames, the UI has no clue how to keep individual frames apart.

Right now I am looking for some feedback and actual performance numbers on both amd and nvidia cards in vertex limited scenes. Some initial testing (on amd) from a few months ago would indicate that this should result in a decent speedup.

If I can get confirmation that this actually improves performance I will try to do the same for the mobile renderer and maybe take a look at the fragment shader.

TODO:
- [ ] get performance numbers from a variety of different gpus and scenes (Help needed!)
- [ ] try to clean things up a bit
- [ ] look in the mobile vertex shader
- [ ] fragment maybe???