@mrjustaguy 2024-01-06
On a 1050ti in my triangle heavy scene, got a fairly consistent 6-7% uplift.

I'd also test on my RX 6600 if it weren't getting RMA-d due to booting issues..