@Calinou 2024-01-16
Tested on a RTX 4090:

Benchmark | `master` 107f2961ccfac179af7682eb5f6e7ea91e80040c | PR rebased against `master` 107f2961ccfac179af7682eb5f6e7ea91e80040c
-|-|-
Platformer 1152×648 | 3415 FPS | 3411 FPS
Platformer 3840×2160 | 941 FPS | 939 FPS
Truck Town 1152×648 | 3428 FPS | 3432 FPS
Truck Town 3840×2160 | 899 FPS | 897 FPS
[Reflection](https://github.com/Calinou/godot-reflection) 1920×1080 | 785 FPS | 779 FPS
[Reflection](https://github.com/Calinou/godot-reflection) 3840×2160 | 252 FPS | 255 FPS

I'll need to check on slower GPUs but so far, I didn't notice a slowdown beyond margin of error at least.