@Ansraer 2024-01-18
Thanks Calinou for testing it. A 4090 is not quite the hardware I had in mind when developing this, but interesting to see nonetheless.
FYI this PR is an optimization primarily meant for large, vertex limited scenes. Looking the scenes you tested I don't think any of them have that high of a vertex count, so sadly I don't think they would show much if a difference.
I myself used blender's subdiv modifier to artificially increase the vertex count of a test scene before creating multiple instances of it inside godot. It's quite an artificial test case but the best I had access to, since most of the free demo scenes out there are quite limited in either size or asset quality.

Now that I am back home I will need to properly test this myself, hopefully I will have some time next weekend.