@Calinou 2024-01-23
Benchmarks on my laptop (Intel Iris Xe, NVIDIA GeForce RTX 4060 Laptop GPU):

### NVIDIA

Benchmark | `master` 107f2961ccfac179af7682eb5f6e7ea91e80040c | PR rebased against `master` 107f2961ccfac179af7682eb5f6e7ea91e80040c
-|-|-
Platformer 1152×648 | 1188 FPS | 1207 FPS
Platformer 2560×1600 | 463 FPS | 458 FPS
Truck Town 1152×648 | 1205 FPS | 1199 FPS
Truck Town 2560×1600 | 473 FPS | 470 FPS
[Reflection](https://github.com/Calinou/godot-reflection) 1920×1080 | 234 FPS | 234 FPS
[Reflection](https://github.com/Calinou/godot-reflection) 2560×1600 | 138 FPS | 136 FPS

### Intel

Benchmark | `master` 107f2961ccfac179af7682eb5f6e7ea91e80040c | PR rebased against `master` 107f2961ccfac179af7682eb5f6e7ea91e80040c
-|-|-
Platformer 1152×648 | 125 FPS | 126 FPS
Platformer 2560×1600 | 34 FPS | 35 FPS
Truck Town 1152×648 | 117 FPS | 117 FPS
Truck Town 2560×1600 | 30 FPS | 30 FPS
[Reflection](https://github.com/Calinou/godot-reflection) 1920×1080 | 24 FPS | 24 FPS
[Reflection](https://github.com/Calinou/godot-reflection) 2560×1600 | 14 FPS | 14 FPS

With https://github.com/Calinou/godot-reflection's `-- --benchmark` command line argument:

## `master`

### NVIDIA

          | Average             | 1% low              | 0.1% low
---------:|---------------------|---------------------|------------------------
Frametime | 210 FPS (4.75 mspf) | 87 FPS (11.47 mspf) | 17 FPS (57.29 mspf)
 CPU Time | 1821 FPS (0.55 mspf) | 1557 FPS (0.64 mspf) | 1053 FPS (0.95 mspf)
 GPU Time | 237 FPS (4.20 mspf) | 161 FPS (6.19 mspf) | 150 FPS (6.66 mspf)

### Intel

          | Average             | 1% low              | 0.1% low
---------:|---------------------|---------------------|------------------------
Frametime | 20 FPS (48.43 mspf) | 14 FPS (67.22 mspf) | 0 FPS (2618.75 mspf)
 CPU Time | 343 FPS (2.91 mspf) | 535 FPS (1.87 mspf) | 0 FPS (2427.16 mspf)
 GPU Time | 24 FPS (40.40 mspf) | 16 FPS (60.86 mspf) | 8 FPS (112.48 mspf)

## This PR

### NVIDIA

          | Average             | 1% low              | 0.1% low
---------:|---------------------|---------------------|------------------------
Frametime | 211 FPS (4.73 mspf) | 85 FPS (11.67 mspf) | 29 FPS (34.10 mspf)
 CPU Time | 2071 FPS (0.48 mspf) | 1934 FPS (0.52 mspf) | 1512 FPS (0.66 mspf)
 GPU Time | 235 FPS (4.25 mspf) | 160 FPS (6.25 mspf) | 151 FPS (6.61 mspf)

### Intel

          | Average             | 1% low              | 0.1% low
---------:|---------------------|---------------------|------------------------
Frametime | 20 FPS (48.06 mspf) | 15 FPS (64.33 mspf) | 0 FPS (2369.34 mspf)
 CPU Time | 364 FPS (2.75 mspf) | 486 FPS (2.06 mspf) | 0 FPS (2193.44 mspf)
 GPU Time | 24 FPS (40.51 mspf) | 16 FPS (59.98 mspf) | 8 FPS (114.15 mspf)

This proves that it's not a GPU-specific issue, it's just that these scenes don't have enough vertices to make a difference.