@Calinou 2024-01-23
That said, I did try creating a scene with lots of vertices (a highly subdivided BoxMesh with 512x512x512 subdivisions), and I still don't notice any performance benefit regardless of resolution on either Intel or NVIDIA GPUs on the laptop mentioned above.

[test_vertex.zip](https://github.com/godotengine/godot/files/14024669/test_vertex.zip)